add `babel-register` wrapper files
