+++ .circleci/configurations/jobs.yml
@@ -95,9 +95,7 @@ jobs:
       # Run JavaScript tests
       - run:
           name: "Run Tests: JavaScript Tests"
-          command: |
-            yarn build
-            node ./scripts/run-ci-javascript-tests.js --maxWorkers 2
+          command: node ./scripts/run-ci-javascript-tests.js --maxWorkers 2
       - run_e2e:
           platform: js
