+++ packages/assets/package.json
@@ -12,14 +12,13 @@
   "keywords": ["assets", "registry", "react-native", "support"],
   "bugs": "https://github.com/facebook/react-native/issues",
   "exports": {
-    "./package.json": "./package.json",
-    "./path-support": "./dist/path-support.js",
-    "./registry": "./dist/registry.js"
+    "./path-support": "./src/path-support.js",
+    "./path-support.js": "./src/path-support.js",
+    "./registry": "./src/registry.js",
+    "./registry.js": "./src/registry.js",
+    "./package.json": "./package.json"
   },
   "files": [
-    "dist",
-    "path-support.js",
-    "registry.js",
-    "src"
+    "dist"
   ]
 }
