+++ packages/assets/registry.js.flow
