+++ packages/assets/src/registry.js
@@ -4,31 +4,17 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @flow strict
+ * @flow
  * @format
+ * @oncall react_native
  */
 
-export type PackagerAsset = {
-  +__packager_asset: boolean,
-  +fileSystemLocation: string,
-  +httpServerLocation: string,
-  +width: ?number,
-  +height: ?number,
-  +scales: Array<number>,
-  +hash: string,
-  +name: string,
-  +type: string,
-  ...
-};
+/*::
+export type * from './registry.flow';
+*/
 
-const assets: Array<PackagerAsset> = [];
-
-export function registerAsset(asset: PackagerAsset): number {
-  // `push` returns new array length, so the first asset will
-  // get id 1 (not 0) to make the value truthy
-  return assets.push(asset);
+if (!process.env.BUILD_EXCLUDE_BABEL_REGISTER) {
+  require('../../../scripts/build/babel-register').registerForMonorepo();
 }
 
-export function getAssetByID(assetId: number): PackagerAsset {
-  return assets[assetId - 1];
-}
+export * from './registry.flow';
